Add domain entities and DTOs for media management; implement enums for tags, formats, collections, and relations

+++ PersonalCollectionShelf.App/ViewModels/LibraryViewModel.cs
@@ -266,7 +266,7 @@ public partial class LibraryViewModel : BaseViewModel, IQueryAttributable
                 return "--";
             }
 
-            return (ratedItems.Sum(item => item.Rating!.Value) / (double)ratedItems.Count).ToString("0.0", CultureInfo.InvariantCulture);
+            return (ratedItems.Sum(item => item.Rating!.Value) / ratedItems.Count).ToString("0.0", CultureInfo.InvariantCulture);
         }
     }
